Worked with codex to try to improve and fix testing for the web application.  Not great results but for now it will do.
